Enhance Stripe webhook handling for subscription updates and deletions. Introduced updateWorkspacePlan utility to streamline plan updates and added logic to check for active subscriptions upon deletion. Cleaned up unused code in customerSubscriptionUpdated.
