+++ apps/web/app/(ee)/api/stripe/webhook/customer-subscription-deleted.ts
@@ -1,14 +1,15 @@
 import { deleteWorkspaceFolders } from "@/lib/api/folders/delete-workspace-folders";
 import { tokenCache } from "@/lib/auth/token-cache";
 import { isBlacklistedEmail } from "@/lib/edge-config/is-blacklisted-email";
+import { stripe } from "@/lib/stripe";
 import { recordLink } from "@/lib/tinybird";
 import { redis } from "@/lib/upstash";
 import { webhookCache } from "@/lib/webhook/cache";
 import { prisma } from "@dub/prisma";
-import { FREE_PLAN, log } from "@dub/utils";
+import { FREE_PLAN, getPlanFromPriceId, log } from "@dub/utils";
 import { NextResponse } from "next/server";
 import Stripe from "stripe";
-import { sendCancellationFeedback } from "./utils";
+import { sendCancellationFeedback, updateWorkspacePlan } from "./utils";
 
 export async function customerSubscriptionDeleted(event: Stripe.Event) {
   const subscriptionDeleted = event.data.object as Stripe.Subscription;
@@ -24,8 +25,10 @@ export async function customerSubscriptionDeleted(event: Stripe.Event) {
     select: {
       id: true,
       slug: true,
-      domains: true,
+      plan: true,
       foldersUsage: true,
+      paymentFailedAt: true,
+      payoutsLimit: true,
       links: {
         where: {
           key: "_root",
@@ -68,6 +71,26 @@ export async function customerSubscriptionDeleted(event: Stripe.Event) {
         stripeId +
         "`* not found in Stripe webhook `customer.subscription.deleted` callback",
     );
+    return NextResponse.json({ received: true });
+  }
+
+  // Check if the customer has another active subscription
+  const { data: activeSubscriptions } = await stripe.subscriptions.list({
+    customer: stripeId,
+    status: "active",
+  });
+
+  if (activeSubscriptions.length > 0) {
+    const activeSubscription = activeSubscriptions[0];
+    const priceId = activeSubscription.items.data[0].price.id;
+    const plan = getPlanFromPriceId(priceId);
+
+    await updateWorkspacePlan({
+      workspace,
+      plan,
+      priceId,
+    });
+
     return NextResponse.json({ received: true });
   }
